requirements.txt
click==7.1.2
Flask==1.1.2
future==0.18.2
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyserial==3.5
PyYAML==5.3.1
Werkzeug==1.0.1
